Added support for configuration files Split examples more granular Migrated modules to eventify package
