+++ eventify/example/__init__.py
@@ -1,0 +1,3 @@
+"""
+Generate RTD for Example
+"""
