+++ eventify/example/config.json
@@ -1,0 +1,8 @@
+{
+    "_comment": "service configuration",
+    "events_to_process": {
+        "NodeProvisionRequested": "provision_node",
+        "NodeStartRequested": "start_node"
+    }
+}
+
