+++ eventify/example/producer.py
@@ -1,5 +1,7 @@
 """
 Example Async Producer
+
+curl -X POST localhost:8080 -d '{"event": "ClientRequested", "body": "im an event! 2", "company_id": 5, "app_id": 1, "global_tenant_id": "1234"}'
 """
 import tornado.ioloop
 import tornado.web
